Fixed climbing code crash, other improvements
Fixed code crashing if next climb setpoint is pressed repeatedly.
Improved shooting software lock by requiring the camera to have a target and robot to be aimed at the goal when doing vision shots. A better method than the way I did likely exists, but this should work.
Added basic code to semi-deploy intake, logic of when to do so is not yet implemented.
Tidied a few code files